first steps towards scheme refactor
