+++ src/schemes/contributionRewardClass.ts
@@ -1,0 +1,10 @@
+import BN = require('bn.js')
+import { Arc } from '../arc'
+import { Proposal } from '../proposal'
+import { ProposalBase } from '../proposalBase'
+import { Address } from '../types'
+import { NULL_ADDRESS } from '../utils'
+
+export class ContributionRewardProposal extends ProposalBase {
+  public state()
+}
